backend/plant_api/requirement.txt
fastapi==0.95.2
uvicorn==0.22.0
python-multipart==0.0.6
torch==2.0.1
torchvision==0.15.2
Pillow==10.0.0